refactor: 목데이터 디렉토리 분리 #249
