+++ frontend/src/mocks/data/statistics/dayStatistics.json
