+++ frontend/src/mocks/data/statistics/weekStatistics.json
